Move ModuleParts to correct folder/namespace
